coresight: Move all sysfs code to sysfs file
BugLink: https://bugs.launchpad.net/bugs/2092141

At the moment the core file contains both sysfs functionality and
core functionality, while the Perf mode is in a separate file in
coresight-etm-perf.c

Many of the functions have ambiguous names like
coresight_enable_source() which actually only work in relation to the
sysfs mode. To avoid further confusion, move everything that isn't core
functionality into the sysfs file and append  _sysfs to the ambiguous
functions.

Link: https://lore.kernel.org/r/20240129154050.569566-7-james.clark@arm.com
(cherry picked from commit 1f5149c7751c50aba1a871143ffa6cb36af3fb49)
Acked-by: Brad Figg <bfigg@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
